requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
amqp==5.3.1
annotated-types==0.7.0
asgiref==3.10.0
attrs==25.4.0
autobahn==25.12.1
Automat==25.4.16
billiard==4.2.2
cachetools==6.2.2
cbor2==5.7.1
celery==5.3.4
certifi==2025.10.5
cffi==2.0.0
channels==4.3.2
channels_redis==4.3.0
charset-normalizer==3.4.4
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
constantly==23.10.4
cryptography==46.0.3
daphne==4.2.1
Django==5.1.2
et_xmlfile==2.0.0
frozenlist==1.8.0
google-ai-generativelanguage==0.6.15
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.43.0
google-auth-httplib2==0.2.1
google-generativeai==0.8.5
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
httplib2==0.31.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
jalali_core==1.0.0
jdatetime==5.2.0
kombu==5.5.4
msgpack==1.1.2
multidict==6.7.0
numpy==1.26.4
openpyxl==3.1.2
packaging==25.0
pandas==2.1.4
prompt_toolkit==3.0.52
propcache==0.4.1
proto-plus==1.26.1
protobuf==5.29.5
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.5
pydantic_core==2.41.5
pyOpenSSL==25.3.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-decouple==3.8
pytz==2025.2
redis==5.0.1
requests==2.31.0
rsa==4.9.1
service-identity==24.2.0
six==1.17.0
sqlparse==0.5.3
tqdm==4.67.1
Twisted==25.5.0
txaio==25.12.2
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
yarl==1.22.0
zope.interface==8.1.1
